requirements.txt
astroid==2.4.2
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
idna==2.10
import==0.0.1
ipinfo==3.0.0
isort==5.6.4
lazy-object-proxy==1.4.3
mariadb==1.0.4
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.11
wrapt==1.12.1
